[Fix] 플레이어 공격력 연결 및 일부 스크립트 이름 수정

+++ Assets/_Scripts/Monster/Monster.cs
@@ -51,7 +51,7 @@ public class Monster : MonoBehaviour
         stateMachine.OnTriggerEnter2D(collision);
     }
 
-    public void TakeDamage(int Damage)
+    public void TakeDamage(float Damage)
     {
         StartCoroutine(ChangeColor());
         AudioManager.Instance.PlaySFX("Boss_Damage");
